// extension/src/components/errorlens/declares.d.ts
export declare interface IELExtensionConfigType {
// ...
  statusBarIconsTargetProblems: TELDiagnosticTarget;
  /**
   * When enabled - highlights status bar icons with background, when disabled - with foreground.
   */
  statusBarIconsUseBackground: boolean;
  /**
   * What to do when there are 0 errors/warnings - hide the item or strip its background color.
   */
  statusBarIconsAtZero: 'hide' | 'removeBackground';
// ...
}

export declare type TELExtensionConfig = Readonly<IELExtensionConfigType>;

/**
 * Common groups of diagnostics to target.
 */
export declare type TELDiagnosticTarget =
  | 'activeEditor'
  | 'all'
  | 'visibleEditors';
// ...
export declare type TELStatusBarProblemType = 'error' | 'warning';

export declare interface IELStatusBarIconsInit {
  isEnabled: boolean;
  atZero: TELExtensionConfig['statusBarIconsAtZero'];
  useBackground: TELExtensionConfig['statusBarIconsUseBackground'];
  targetProblems: TELExtensionConfig['statusBarIconsTargetProblems'];
}